backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.8
Pillow>=9.0.0
pandas==2.1.4
requests==2.31.0